pyproject.toml
[project]
name = "decentralized-forum"
version = "0.1.0"
description = "A decentralized autonomous forum with gamification and blockchain rewards"
authors = [{name = "Forum Creator"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.3.2",
    "httpx>=0.27.2",
    "web3>=7.5.0",
    "eth-account>=0.13.0",
    "eth-typing>=5.0.0",
    "pillow>=11.0.0",
    "markdown>=3.7",
    "bleach>=6.2.0",
    "redis>=5.2.0",
    "celery>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.7.3",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.7.3",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
